pyproject.toml
[project]
name = "ev-load-fc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "catboost>=1.2.8",
    "cupy-cuda12x>=13.6.0",
    "duckdb>=1.4.1",
    "ipykernel>=7.2.0",
    "ipython>=9.10.0",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.7",
    "meteostat>=1.7.6",
    "mlflow>=3.8.0",
    "nbformat>=5.10.4",
    "numpy>=2.3.3",
    "optuna>=4.6.0",
    "pandas>=2.3.3",
    "pathlib>=1.0.1",
    "pmdarima>=2.1.0",
    "prophet>=1.2.1",
    "pytest>=9.0.2",
    "seaborn>=0.13.2",
    "shap>=0.49.1",
    "statsmodels>=0.14.5",
    "word2number>=1.1",
    "xgboost-cu12==3.1.2",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# Treat project root as the base and explicitly search each src root.
package-dir = { "" = "." }

[tool.setuptools.packages.find] 
where = ["src"] 
include = ["ev_load_fc*"]

# --------------------
# mypy configuration
# --------------------
[tool.mypy]
python_version = "3.11"
strict = false
mypy_path = "src/"
namespace_packages = true
explicit_package_bases = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_return_any = false
warn_unreachable = true
disallow_any_unimported = false
disallow_any_expr = false
disallow_any_generics = false
disallow_any_decorated = false
no_implicit_optional = true
check_untyped_defs = false

exclude = [
    "tests/",
]


# --------------------
# pytest (optional)
# --------------------
[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "bandit>=1.9.3",
    "mypy>=1.19.1",
    "pip-licenses>=5.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
    "ty>=0.0.16",
]
